requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
APScheduler==3.10.4
openpyxl==3.1.5
